+++ snuba/query/data_source/simple.py
@@ -81,7 +81,9 @@ class Storage(LogicalDataSource):
         """There should be no operations done on a Storage except to get the storage key
         Therefore the columns are empty
         """
-        raise NotImplementedError()
+        raise NotImplementedError(
+            "Storage queries do not support entity processing or Joins"
+        )
 
 
 @dataclass(frozen=True)
